reviews_service/requirements.txt
flask
flask-cors
SQLAlchemy
psycopg2-binary
pyjwt
gunicorn
